api/requirements.txt
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
decorator
exceptiongroup
fastapi
fastcore
fasteners
fastjsonschema
filelock
gitdb
GitPython
h11
idna
importlib-metadata
Jinja2
json5
jsonpointer
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
mpmath
networkx
numpy
opencv-python
opencv-python-headless
packaging
pandas
Pillow
protobuf
pyarrow
pydantic
pydantic_core
pydeck
Pygments
Pympler
python-dateutil
python-multipart
pytz
pytz-deprecation-shim
referencing
requests
rich
rpds-py
six
smmap
sniffio
starlette
streamlit
sympy
tenacity
toml
toolz
torch
torchaudio
torchvision
tornado
transformers
typing_extensions
tzdata
tzlocal
ultralytics
urllib3
uvicorn
validators
zipp
